refactor(api): move stuff around
