+++ apps/api/src/app/tests/request-parser/api.request-parser.spec.ts
@@ -2,7 +2,7 @@ import { INestApplication } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import * as request from 'supertest';
 
-import { RequestParserTestController } from '../controllers/test.request-parser.controller';
+import { RequestParserTestController } from './test.request-parser.controller';
 
 describe('API Request Parser', () => {
   let app: INestApplication;
